Update README.md
This repository provides a comprehensive list of all the Auto Assembler commands used in Cheat Engine, a powerful open-source memory scanner and debugger. Auto Assembler allows users to write custom scripts to manipulate and interact with the memory of a target process, making it a popular choice for game hacking and software reverse engineering tasks.